requirements.txt
streamlit 
langchain
langchain-groq
langchain_huggingface 
langchain_community
python-dotenv
langserve 
fastapi
uvicorn
sse_starlette
-e .